reorganized utils functions
