+++ README.md
@@ -70,7 +70,7 @@ The available calls within an executor are:
 |[map_reduce()](docs/api-details.md#executormap_reduce) | Async. | Method used to spawn multiple function activations with one (or multiple) reducers|
 |[wait()](docs/api-details.md#executorwait) | Sync. | Wait for the function activations to complete. It blocks the local execution until all the function activations finished their execution (configurable)|
 |[get_result()](docs/api-details.md#executorget_result) | Sync. | Method used to retrieve the results of all function activations. The results are returned within an ordered list, where each element of the list is the result of one activation|
-|[plot()](docs/api-details.md#plot) | Sync. | Method used to create execution plots |
+|[plot()](docs/api-details.md#executorplot) | Sync. | Method used to create execution plots |
 |[clean()](docs/api-details.md#executorclean) | Async. | Method used to clean the temporary data generated by PyWren in IBM COS |
 
 Additional information and examples:
